use proper initializer for structs/arrays
one i was using is a c23 extension i think. could just not use initializer at all but that doens't zero out the struct/array

according to this link, this is a valid way to 0 out a block scope array in c

https://en.cppreference.com/w/c/language/array_initialization

according to this: https://stackoverflow.com/a/6891773

this is in the standard for structs as well?