revert to checking by byte instead of by long
checking header by long assumed the header would always be aligned to 4 bytes which does seem to be the case for sm3 files but checking by byte is more robust.

only issue is this is likely a lot less performant. as it assigns the index variable in a loop when the header is not matching

partially reverts 6696949a