dont malloc result array in printFSBHeaderIndexes
undoes what 09726e06 did

i didnt realise this before but this isn't actually needed. reason i thought it did something before was without it i was printing 0 and with it i was printing some big number. but that was just cause malloc gave some uniniitalized data while before it was initialized to 0. malloc didn't actually help, the problem was something else.

now that i've fixed the problem, it is safer to avoid using malloc. in my opinion.