attempt append
tried doing this code block thing to make it easier to see the file handle uses, but it's still difficult to read.

mainly because im keeping the filehandles open so i can use them later. it might be easier to just close the file immediately and then reopen when i need to use it again