use block scoping for malloc and file handles
kinda discovered this in last commit, but i think this makes the code a lot cleaner. cause its similar to autoclose or using in other languages except you still have to close the resource yourself. still it helps organise.

i had to move a few variables around because they were used inside and outside the block scope but it worked out nicely.

searched this up and i saw someone else have this idea on stackoverflow too

https://stackoverflow.com/a/1680664