fix 0s at the end of the file.
the readAndAppend function was appending readCount bytes to the output even when not that many were read from the input. i forgot to use the numRead output of fread to only write what has been read.

with this fix most of the replaceAudio function seems to work well. i checked the output file and I didn't see any missing bytes or anything where the replace audio was put.

the writing data size part doesn't work though, leads to a file with just zeroes in it. will take a look at that next.