* fix implicit conversion changes signedness warning
to safely cast i needed to make sure the index wasn't negative. only case i could think of is when fileextensionptr is null, so i added a check for that.

* fix up some fprintfs

forgot to close the file on one and also i had put log instead of error when it was an error