generate output file name using input file name with "-mod" at the end.
was thinking of doing it a different way by having the output file be in the same folder as the executable or something. but that would be a lot harder to do because i'd need to extract the filename from pcssbFileName without the path before it. on posix that's easy but doing a cross platform function would be tedious in c. i think such a function is in c++... im seriously considering switching the codebase to that after i finish this functionality.

but anyway, doing this with snprintf is a lot easier. I was thinking of using sprintf but it's not safe... neither is strcpy or strncpy, i might switch all those string building parts of the code to use snprintf.

most relevant:
https://stackoverflow.com/a/12276028

---

https://stackoverflow.com/a/9937228

https://stackoverflow.com/a/14905963

https://en.cppreference.com/w/c/string/byte/strlen

https://en.cppreference.com/w/c/io/fprintf