fix memory issue in outputAudioFiles (fixes problem when outputting soundtrack_common.pcssb)
one of the file names in soundtrack_common.pcssb ("City_Black_Music_Variation_1.w") took the full 30 bytes and didn't have any 00 bytes.

when this file name was read there was nothing to null terminate it, causing the overflow when snprintf was called in outputAudioFiles, as it was trying to read out of bounds.

for simplicity I just added one to the FSB_FILENAME_SIZE macro and added a line to add a null terminator in readFileName.