change readAndAppend to readAndWriteToNewFile
now we open the file in write mode for the first operation. ensures the file is started from scratch effectively, and that we don't add on to an existing file. however it means that whatever file was there before gets overwritten.

was in a bit of a dilemma on how to assign outputFileHandle, because I wouldn't be able to do it in an if else block checking the append bool while keeping it const. luckily ternary operator lets us keep it const though