accidentallly commited myIO changes in 85c3c74b. so i was expanding on those. basically there I exempted EEXIST from the error printing because itw as being spammed in the output when you run the program.
what I didn't realise was that was because the directory creation was in the for loop so mkdir was being run for every fsb file within the pcssb, which is obviously unnecessary. it only needs to be run once. now that EEXIST exemption is not needed because it only shows once

i also added pcssb and sm3tools back to gitignore. i have a barebones cmake that builds to bin but for now im still usign my original makefile which builds in the root directory