add malloc and calloc null checks
only buffer used with calloc had a warning for it potentially being null but im guessing that was just because we wrote directly to an index in the buffer.

was thinking of making a mymalloc function but i dont think its used enough times to warrant it. especially when the check is only few lines.