use "MyIO" namespace for myIO functions
i wasnt sure about whether it would be worth doing this, but I guess since this myIO is kind of a separate thing to the main project it makes sense to put it into a name space. i had all the functions prefixed with "my" but this is just the proper way of avoiding name collisions

though with that logic should i not do a namespace for pcssb? or should i do one?