use args.size() instead of argc for for loop
this fixes the "comparison of integers of different signs" warning

there are loads of other ways we could have fixed this.

one example would be
for (auto i = args.begin(); i != args.end(); ++i)

 and then use *args instead of args[i].

another example would be

for (const std::string& arg : args)

in which case i'm using a const reference to prevent copying.

ultimately i decided on sticking with the regular for loop but using args.size() to prevent the sign warning