return std::string from readFileName instead of writing to a resultArr
i think this is a LOT safer because we avoid pointer decay and such things.

and it's just cleaner to actually return something rather than have the function write to a function parameter.

i could have done some things to have fread read into the std::string directly, but i thought it would be better to use a std::array for the buffer.

 https://stackoverflow.com/questions/13612340/c-fread-into-a-stdstring

 https://www.learncpp.com/cpp-tutorial/stdstring_view-part-2/