Revert "hmm.... ok i was thinking off with starting with the getfileextension from inputfilepath  first and then potentially changing it from there... but how do i deal with the case where the file extension isn't recognised? do i add an "other" file type that is returned from that function in that case?"
Decided to not have the file type flag after all. I don't think it's needed, as there wouldn't be a case where you'd want a file that is one thing treated as if it was another, it would just lead to breakage. Program isn't designed for that to work.

The reason why I wanted such an option was to handle files that had extensions like "pcssb.MOD" but it would be better served detecting the file type from the magic number in the file itself. E.g. "FSB3" for PCSSB

This reverts commit 75a91e7a83aee7ac897c55692d867791eededc25.