add findflagindexes function
storing indexes of flags will help with e.g. default arg position for if the --input flag is not specified. right now that's broken as it will just take the first argument which could be a flag such as --list

i don't really like the findInputFileArg function as it isn't reusable, will see if i can get rid of it.