add checkFlagPresent and getFlagArgument functions
inspired by https://medium.com/@mostsignificant/3-ways-to-parse-command-line-arguments-in-c-quick-do-it-yourself-or-comprehensive-36913284460f

significantly simplifies structure of parseflags.

only problem is i don't know how easy it would be to check the flag argument with in getFlagArgument, because it doesn't use a list of flags. also for the input file i want it to default to the first non flag argument if the input flag is not passed and that is the most difficult thing to do