Fix filenames not matching in findFirstFSBMatchingFileName()
The issue with matching the filename was because I was specifying the length when constructing the std::string, and set it to be 30. I assumed that acted as "maximum length" but apparently it treats the input as a char array rather than a c string. I.E. it won't stop when it encounters the \0. So the length of the string ended up being different even if it looked the same.

see https://codeberg.org/SpiderGlider/sm3tools/issues/17#issuecomment-5634818

https://stackoverflow.com/questions/15490690/c-comparing-a-reference-string-to-a-string

https://stackoverflow.com/a/164274