error when inputFilePath is empty
I had found that when calling the program with the input file as an argument (without the --input flag), then it would say "ERROR: Program needs an input file" due to how the program is structured. This still gives an error but at least it's a slightly more fitting one. But I still want to try to make it so that in this scenario it does actually take the file as an argument