document arg functions in sm3tools.hpp
renamed getFlagArgument to getFlagValue because can easily be confused with findArgument. am now distinguishing between "flag value" meaning the argument to a flag and "argument" which is just a non flag argument passed to the program.

renamed getArgument to getArgOrFlagValue for similar reasons.

with that done, i now rename findArgument to getArgument. it was only "find" before as a previous attempt to distinguish the function names but now it's been done in a better way

took inspiration for some phrases from here https://clig.dev/#arguments-and-flags