cleanup printFSBList
static casted the data size to unsigned long so i could use the lu format specifier. this shouldn't result in any loss of data.

there is a portable format specifier for uint32_t but I would have to include another header just for this and i'd also have to define something for c++ apparently.

https://stackoverflow.com/a/9225648