move usage text into a constexpr variable
unlike the other constexprs, i've kept it in the cpp file rather than the hpp. I just think it make a lot more sense to be defined in the printHelp function.

the function would be kind of useless if hte text was defined outside of it, as it's just a simple cout...