add flags to usage text
not sure if there's a better way to indent using cout, but this seems ok for now

the text is mainly copied from the readme, but i altered the ordering of one of the flags so the flags without args appear at the bottom. makes the | separator look nicer.

formatting inspired a little by manpages